% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteNumbers.R
\name{splitIntoNumbers}
\alias{splitIntoNumbers}
\title{Extract number strings from a character vector excluding commas}
\usage{
splitIntoNumbers(x, end_pos = -1L)
}
\arguments{
\item{x}{An character vector (which must not contain NA and NULL) or a NULL or a NA.}

\item{end_pos}{The position of the last character of concatenated x.
-1 means using fully concatenated x without trimming.}
}
\value{
The character which contains space-separated numbers in x.
}
\description{
This function removes commas from an input character vector,
  treats consecutive non-numeric characters of them as delimiters,
  and extracts number strings from them.
}
\examples{
splitIntoNumbers("3")
splitIntoNumbers("3.")
splitIntoNumbers("3.25")
splitIntoNumbers(".25")
splitIntoNumbers("1,234")
splitIntoNumbers(x="1,234", end_pos=4)
splitIntoNumbers("1 -2.5")
splitIntoNumbers(NULL)
splitIntoNumbers(NA)
splitIntoNumbers(character())
splitIntoNumbers(c("1.", "2", "3.5"))
}
